Moved the included file of 27.c in the include directory.
